Fixed compile errors for ImageTools when compiling for x64
Binaries submitted in 770837
Build time : 0 minutes 5.179000 seconds
Branch : //eve/branches/development/MAINLINE/

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/ImageTools/": change = 770836]

+++ FileStream.cpp
@@ -77,7 +77,7 @@ ptrdiff_t FileStream::Seek( ptrdiff_t distance, SeekOrigin method )
 		default:
 			return -1;
 		}
-		if( fseek( m_file, distance, origin ) )
+		if( fseek( m_file, long( distance ), origin ) )
 		{
 			return -1;
 		}
